+++ tcex/tcex/testing/test_case_playbook.py
@@ -24,27 +24,34 @@ class TestCasePlaybook(TestCasePlaybookCommon):
 
         args['tc_playbook_out_variables'] = ','.join(self.output_variables)
         self.log_data('run', 'args', args)
-        app = self.app(args)
+        self.app = self.app_init(args)
 
         # Start
-        exit_code = self.run_app_method(app, 'start')
+        exit_code = self.run_app_method(self.app, 'start')
         if exit_code != 0:
             return exit_code
 
         # Run
         try:
-            if hasattr(app.args, 'tc_action') and app.args.tc_action is not None:
-                tc_action = app.args.tc_action
+            if hasattr(self.app.args, 'tc_action') and self.app.args.tc_action is not None:
+                tc_action = self.app.args.tc_action
                 tc_action_formatted = tc_action.lower().replace(' ', '_')
                 tc_action_map = 'tc_action_map'
-                if hasattr(app, tc_action):
-                    getattr(app, tc_action)()
-                elif hasattr(app, tc_action_formatted):
-                    getattr(app, tc_action_formatted)()
-                elif hasattr(app, tc_action_map):
-                    app.tc_action_map.get(app.args.tc_action)()  # pylint: disable=no-member
+                if hasattr(self.app, tc_action):
+                    getattr(self.app, tc_action)()
+                elif hasattr(self.app, tc_action_formatted):
+                    getattr(self.app, tc_action_formatted)()
+                elif hasattr(self.app, tc_action_map):
+                    self.app.tc_action_map.get(
+                        self.app.args.tc_action
+                    )()  # pylint: disable=no-member
+                else:
+                    self.log.error(
+                        'Action method ({}) was not found.'.format(self.app.args.tc_action)
+                    )
+                    self._exit(1)
             else:
-                app.run()
+                self.app.run()
         except SystemExit as e:
             self.log.error('App failed in run() method ({}).'.format(e))
             return self._exit(e.code)
@@ -55,17 +62,22 @@ class TestCasePlaybook(TestCasePlaybookCommon):
             return self._exit(1)
 
         # Write Output
-        exit_code = self.run_app_method(app, 'write_output')
+        exit_code = self.run_app_method(self.app, 'write_output')
         if exit_code != 0:
             return exit_code
 
         # Done
-        exit_code = self.run_app_method(app, 'done')
+        exit_code = self.run_app_method(self.app, 'done')
         if exit_code != 0:
             return exit_code
 
-        app.tcex.log.info('Exit Code: {}'.format(app.tcex.exit_code))
-        return self._exit(app.tcex.exit_code)
+        try:
+            # call exit for message_tc output, but don't exit
+            self.app.tcex.playbook.exit(msg=self.app.exit_message)
+        except SystemExit:
+            pass
+
+        return self._exit(self.app.tcex.exit_code)
 
     def run_profile(self, profile):
         """Run an App using the profile name."""
@@ -80,9 +92,8 @@ class TestCasePlaybook(TestCasePlaybookCommon):
         # run the App
         exit_code = self.run(args)
 
-        # populate the output variables
-        if exit_code == 0:
-            self.populate_output_variables(profile)
+        # add context for populating output variables
+        self.context_tracker.append(self.context)
 
         return exit_code
 
@@ -91,10 +102,3 @@ class TestCasePlaybook(TestCasePlaybookCommon):
         super(TestCasePlaybook, self).setup_method()
         self.stager.redis.from_dict(self.redis_staging_data)
         self.redis_client = self.tcex.playbook.db.r
-
-    def teardown_method(self):
-        """Run after each test method runs."""
-        r = self.stager.redis.delete_context(self.context)
-        self.stager.threatconnect.delete_staged(self._staged_tc_data)
-        self.log_data('teardown method', 'delete count', r)
-        super(TestCasePlaybook, self).teardown_method()