+++ docs/tcex/release_notes.rst
@@ -3,6 +3,36 @@
 Release Notes
 #############
 
+1.1.7
+=====
++ Updates to testing framework for custom validation.
++ Updates to the docs for multiple modules.
++ Multiple updates to testing framework.
+
+1.1.6
+=====
++ Updated deleted() method of TI module to yield results instead of returning raw response.
++ Updates to testing framework for custom methods when testing profiles.
++ Updated inputs to ensure args provided via sys.argv take precedent over all other args.
++ Added new service_id arg for service Apps.
++ Added POC of session_external. Python requests session with auto-proxy configuration.
++ Updated excludes for tcpackage command for pytest report folders.
+
+1.1.5
+=====
++ Updated validation module to handle local imports and shared modules.
+
+1.1.4
+=====
++ Added additional support for v2 API endpoints.
++ Added support for new appId field in the install.json.
++ Updated validation command to better handle packages with nested modules.
++ Updated PB module to handle execution with no requested output variables.
++ Updated PB module to handle null values in BinaryArray.
++ Updated TI modules to better handle conversion to and from TCEntity.
++ Updated external App template to allow passing configuration in on TcEx() initialization.
++ Multiple updates for testing framework.
+
 1.1.3
 =====
 + Added cache handler to logging module.